[Web] Refactor the profile folder
