+++ core/datacap-web/console-fe/src/views/user/profile/ProfileGenerate.ts
