+++ core/datacap-web/console-fe/src/views/user/profile/layout/ProfileLayout.vue
@@ -14,8 +14,8 @@
 </template>
 
 <script lang="ts">
-import ProfileLayoutSidebar from "@/views/pages/profile/layout/components/ProfileLayoutSidebar.vue";
-import ProfileLayoutContent from "@/views/pages/profile/layout/components/ProfileLayoutContent.vue";
+import ProfileLayoutSidebar from "@/views/user/profile/layout/components/ProfileLayoutSidebar.vue";
+import ProfileLayoutContent from "@/views/user/profile/layout/components/ProfileLayoutContent.vue";
 
 export default {
   name: "ProfileLayout",
