+++ core/datacap-web/console-fe/src/views/user/profile/layout/components/ProfileLayoutContent.vue
